Move all accessibility-related stuff in one subfolder and namespace
