+++ ReactWindows/ReactNative.Shared/Views/Control/ReactControl.cs
@@ -1,5 +1,6 @@
 using ReactNative.UIManager;
 #if WINDOWS_UWP
+using ReactNative.Accessibility;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Automation.Peers;
 using Windows.UI.Xaml.Controls;
@@ -13,7 +14,11 @@ namespace ReactNative.Views.ControlView
     /// <summary>
     /// A native control with a single Canvas child.
     /// </summary>
+#if WINDOWS_UWP
     public class ReactControl : UserControl, IAccessible
+#else
+    public class ReactControl : UserControl
+#endif
     {
         private readonly Canvas _canvas;
 
@@ -61,16 +66,16 @@ namespace ReactNative.Views.ControlView
             set;
         }
 
-        // TODO: implement runtime change raising event to screen reader #1562
-        /// <inheritdoc />                                                    
-        public AccessibilityTrait[] AccessibilityTraits { get; set; }
-
 #if WINDOWS_UWP
         /// <inheritdoc />                                              
         protected override AutomationPeer OnCreateAutomationPeer()
         {
             return new DynamicAutomationPeer<ReactControl>(this);
         }
+
+        // TODO: implement runtime change raising event to screen reader #1562
+        /// <inheritdoc />                                                    
+        public AccessibilityTrait[] AccessibilityTraits { get; set; }
 #endif
     }
 }
