+++ ReactWindows/ReactNative/Views/Text/ReactTextViewManager.cs
@@ -1,4 +1,4 @@
-using ReactNative.Common;
+using ReactNative.Accessibility;
 using ReactNative.Reflection;
 using ReactNative.UIManager;
 using ReactNative.UIManager.Annotations;
